+++ src/main/java/com/esosa/pass_manager/controllers/interfaces/IAuthController.java
@@ -1,0 +1,22 @@
+package com.esosa.pass_manager.controllers.interfaces;
+
+import com.esosa.pass_manager.controllers.request.AuthRequest;
+import com.esosa.pass_manager.controllers.response.AuthResponse;
+import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@RequestMapping("/auth")
+public interface IAuthController {
+
+    @PostMapping("/register")
+    @ResponseStatus(HttpStatus.CREATED)
+    void register(@RequestBody @Valid AuthRequest authRequest);
+
+    @PostMapping("/login")
+    @ResponseStatus(HttpStatus.CREATED)
+    AuthResponse login(@RequestBody @Valid AuthRequest authRequest);
+}
